+++ src/bin/df_test.rs
@@ -1,18 +1,20 @@
 extern crate peroxide;
 use peroxide::*;
-use self::Header::*;
+use std::error::Error;
 
-fn main() {
-    let mut a: DataFrame<Header> = DataFrame::new();
-    a.insert(X, seq(0, 100, 1));
-    a.insert(Y, c!(4,5,6));
-    a.insert(Z, c!(7,8,9));
-    a.print();
-}
+fn main() -> Result<(), Box<dyn Error>> {
+    #[cfg(feature = "dataframe")]
+    {
+        let mut a: DataFrame = DataFrame::new();
+        a.insert("x", c!(1,2,3,4,5));
+        a.insert("y", c!(4,5,6));
+        a.insert("z", c!(7,8,9));
+        a.print();
+        a.write_csv("example_data/df_test.csv")?;
+        a.write_cdf("example_data/df_test.cdf")?;
 
-#[derive(Debug, Hash, PartialEq, Eq, Clone)]
-pub enum Header {
-    X,
-    Y,
-    Z
+        let b = DataFrame::read_csv("example_data/df_test.csv", ',')?;
+        b.print();
+    }
+    Ok(())
 }